move NewSidebar into Sidebars directory
