+++ packages/studio-base/src/components/Sidebars/NewSidebar.tsx
